src/auth/gRPC/requirements.txt
grpc-interceptor ~= 0.12.0
grpcio-tools ~= 1.30
pytest ~= 5.4
SQLAlchemy==1.4.15
psycopg2-binary==2.8.6
bcrypt==3.2.0
redis==3.5.3
PyJWT==2.1.0
pydantic==1.8.1
loguru==0.5.3
ipython